Class hierarchy for agents.
Entity represent tasks that must be performed at every frame.
Agent is base class for objects with reasoning and decision-making capabilities.
Person is a human agent with a transport-mode.
Transport_mode is the base class for the methods a person can employ to move around.
Pedestrian is the walking mode.
Driver is the driving mode.
Cyclist is the cycling mode.
Passenger is the riding mode (on a bus, taxi, train, or private car).