Removed Buffered<T>::is_dirty_ member.
In testing/data/data.hpp, the double-buffered data type is the subject
of the Observer design pattern.  The (double-buffered) object has a list
of observers which will be notified whenever its value was changed.  The
is_dirty_ member determines whether the notify() is invoked.  Since the
Buffered<T> class is not participating in the Observer design pattern,
the is_dirty_ member is not needed.