add class Signal(including .hpp and .cpp file in the folder:entities)
NOTES:
1.The signal light uses adaptive signal control strategy (SCATS).

2.The input value for SCATS is the degree of saturation(DS),this value
can be recorded by loop detector, for now the loop section is
nothing.We can define the value of DS as default or random. 

3.I am not so sure about how to add class Signal into the Basic Project,
such as which folder to put it in, entities or roles? It needs to be
updated and every agent should be able to get its information.