It is now possible to create more than one Signal object, no longer a singleton.
There are still 2 global variables -- Density[] and DS_all.  I believe
Density[] will be modified by the loop-detectors and so should be a data
member of the loop-detector class.  And I think DS_all should be a local
variable inside Signal::updateSignal().  Both variables are not modified
and so can be "shared" by various Signal objects.  Will need to remove
them in future.