Changed code to throw if Signal is called for the wrong Lane or Crossing instead of returning a constant Red color.
Previously, if the modellers calls the wrong Signal object for the
traffic lights for a particular Lane object or Crossing object, the
Signal object returns Red.  The modellers will wonder why the color
doesn't change to Green or Yellow.  Now, we throw a string which
contains information about the Signal and the Lane/Crossing.  Hopefully
the modeller may realize from the thrown message that the Lane/Crossing
are not under the control of the Signal object.  Most likely the
modeller should be using the Signal object that is allocated at the
adjacent intersection.