Added code to render the arrow marking icons.
The arrows are pointing in the correct direction for those that are
pointing in the north-east and south-west directions (eg. along
Victoria Street, Queen Street and North Bridge Road).  However those
that are pointing in the north-west and south-east directions (eg.
along Middle Road, Stamford Road, Bras Basah Road) are pointing in the
opposite directions.  For example, Bras Basah Road is a one-way road and
the arrows are pointing the wrong way.  Middle Road is 2-way, but the
arrow markings are also pointing in the wrong directions.

For arrows that are pointing in the correct direction, their bearings
are not parallel to the lane markings.

Arrows in round-abouts are pointing incorrectly as well.

I don't whether if the errors are caused by a bug in the code that loads
the bearing data from the shape files or whether the bearing data
themselves are errorenous.