Added code to avoid creating Signal objects at nodes with less than 4 links.
Signal.cpp is designed for intersections with 4-way traffic currently.
This needs to be fixed.  For the moment, we do not create Signal object at
nodes with less than 4 links.

This means we will not have Signal objects at uni-nodes; the purpose of such
signals is to allow pedestrians to cross the roads.

We also will not have Signal objects at nodes where some links are not
loaded in, such nodes will be at the edges of the partially-loaded road network.