Changed type of Node::location from Point2D* to const Point2D.
It is really irritating that Node::location is a pointer; so much code
needs to know the node's position and they have to deference the
pointer.  This change makes it more saner and makes the code so much
more readable.

At the same time, I also took the opportunity to make use of
Point2D::operator==() and Point2D copy constructor.  That was so much
un-readable code in the form
    if (start->location.getX() == point.getX() and
        start->location.getY() == point.getY())
when
    if (start->location == point)
is so much more readable.  The latter form uses Point2D::operator==().
There is also the related Point2D::operator!=() method.

There are also code of the form
    Point2D point(start->location.getX(), start->location.getY());
which can be simplified to
    Point2D point(start->location);
The Point2D copy constructor is automatically synthesized bt the
compiler.