Changed controller.py to use Qt Designer.
I foresee that controller.py code size will grow when the missing
functionality are implemented.  I feel that the GUI code will be
distracting when reading the code.  Now the code is cleaner because
the UI forms are in controller.ui.  Modifying is also easier with
Qt Designer.

Please see README on what you need to do before you can use the
visualizer.  The short version is: you need to run ``make'' once.