Tested the communication using 1500 agents and 600 android agents+NS3
refining the agent subscription list is not done anymore as it seems not longer needed.
Still:
1-The 2 minutes Simulation took almost 1+ hour to complete.
visual observation suggests that NS3 had a heavy job to complete, so the
broker had to wait for NS3 completion at each tick.

2-At some early point(~tick 133) the three client process(each running 200
threads) stoped updating their console. The reason is not verified. it
should be checked against the number of registered clients in
simmobility who are still active.
3- eclipse messes with my large amount of output to console. so does the
ubuntu command prompt. may be if we limit the console output size, this
problem would be removed.